Add dedicated 404 page
Created 404.tsx in /pages which is a copy of [...not-found].tsx, and updated the redirect in pages/[id]/index.tsx to redirect to the 404 page if no data is returned from getApplicationById.
I think it should be possible to use the getStaticProps notFound functionality: https://nextjs.org/docs/pages/api-reference/functions/get-static-props#notfound (I.e. return {notFound: true} from getStaticProps, but I can't get it to work.